Rename IngestTaskStatusTestData
